Changed parameter of moveTilesForward to float

+++ mr_roboto_lab_2/mr_roboto_lab_2.ino
@@ -667,7 +667,7 @@ void stopRobot() {
 *
 *
 *************************************************************************/
-void moveTilesForward(int numTiles) {
+void moveTilesForward(float numTiles) {
   moveTicksForward(numTiles * TICKS_PER_TILE);
 }
